requirements.txt
flask
SQLAlchemy==1.4.46
pandas
openai
pandasql
python-dotenv
